Fix crash when obtaining package name
In HockeySDK_Portable\HockeyClient.cs if the AppPackageName is not set a
NullReferenceException is thrown by the getter as
Assembly.GetExecutingAssembly().EntryPoint returns null.

Rather than fixing the issue within the affected property an overload of
the Configure method in HockeyClientWinFormsExtensions has been
implemented whereby the caller can supply the package name.