Error in writing crash data to user's system
The Stream.CopyToAsync method did not seem to copy the contents of the
stream to a file stream object, thus no data was written to the generated
crash log file on the user's system. Have now used the CopyTo method
instead, as well as the Flush method on the file stream object to ensure
that the crash log file has data written to it.